Split up parser building for running docker
