+++ run_docker.py
@@ -20,6 +20,11 @@ It is used to run the main calls to the pipeline, namely:
 """
 import argparse
 import os
+from fetal_brain_qc.cli.build_run_parsers import (
+    build_reports_parser,
+    build_inference_parser,
+)
+from fetal_brain_qc.definitions import BRAIN_CKPT
 
 MASK_PATTERN = (
     "sub-{subject}[/ses-{session}][/{datatype}]/sub-{subject}"
@@ -27,184 +32,21 @@ MASK_PATTERN = (
 )
 
 
-def build_reports_parser(parser):
-    """
-    Build parser for the reports pipeline.
-    """
-    parser.add_argument(
-        "--bids_dir",
-        required=True,
-        help="BIDS directory containing the LR series.",
-    )
-
-    parser.add_argument(
-        "--masks_dir",
-        help="Root of the BIDS directory where brain masks will be stored.",
-        required=True,
-    )
-
-    parser.add_argument(
-        "--reports_dir",
-        help="Directory where the reports will be stored.",
-        required=True,
-    )
-
-    parser.add_argument(
-        "--ckpt_path",
-        help="Path to the checkpoint of the MONAIfbs model.",
-        default=None,
-    )
-
-    parser.add_argument(
-        "--mask_pattern",
-        help=(
-            "Pattern according to which the masks will be stored.\n "
-            'By default, masks will be stored in "<masks_dir>/sub-{subject}[/ses-{session}][/{datatype}]/sub-{subject}'
-            '[_ses-{session}][_acq-{acquisition}][_run-{run}]_{suffix}.nii.gz", and the different fields will be '
-            "substituted based on the structure of bids_dir."
-        ),
-        type=str,
-        default=MASK_PATTERN,
-    )
-    parser.add_argument(
-        "--bids_csv",
-        help="CSV file where the list of available LR series and masks will be stored.",
-        default="bids_csv.csv",
-    )
-
-    parser.add_argument(
-        "--seed",
-        type=int,
-        default=42,
-        help="Seed for the random number generator.",
-    )
-
-
-def build_inference_parser(parser):
-    """
-    Build parser for the inference pipeline.
-    """
-    parser.add_argument(
-        "--bids_dir",
-        required=True,
-        help="BIDS directory containing the LR series.",
-    )
-
-    parser.add_argument(
-        "--masks_dir",
-        help=(
-            "Root of the BIDS directory where brain masks will be/are stored. "
-            "If masks already exist, they will be used."
-        ),
-        required=True,
-    )
-
-    parser.add_argument(
-        "--seg_dir",
-        help=(
-            "Root of the directory where brain segmentations will be stored. "
-            "If segmentations already exist, they will be used."
-        ),
-        required=True,
-    )
-    parser.add_argument(
-        "--bids_csv",
-        help="CSV file where the list of available LR series and masks will be stored.",
-        default="bids_csv.csv",
-    )
-    parser.add_argument(
-        "--iqms_csv",
-        help="CSV file where the computed IQMs will be stored.",
-        default="iqms_csv.csv",
-    )
-
-    parser.add_argument(
-        "--out_csv",
-        help="CSV file where the predictions from FetMRQC will be stored.",
-        default="out_csv.csv",
-    )
-
-    parser.add_argument(
-        "--iqms",
-        help="List of IQMs that will be computed",
-        nargs="+",
-        default="all",
-    )
-
-    parser.add_argument(
-        "--fetmrqc20_iqms",
-        help="Whether the IQMs from FetMRQC-20 should be computed",
-        action=argparse.BooleanOptionalAction,
-        default=True,
-    )
-
-    parser.add_argument(
-        "--use_all_iqms",
-        help="Whether all IQMs should be computed",
-        default=False,
-        action="store_false",
-        dest="fetmrqc20_iqms",
-    )
-    parser.add_argument(
-        "--ckpt_path",
-        help="Path to the checkpoint of the MONAIfbs model.",
-        default=None,
-    )
-
-    parser.add_argument(
-        "--mask_pattern",
-        help=(
-            "Pattern according to which the masks will be stored.\n "
-            'By default, masks will be stored in "<masks_dir>/sub-{subject}[/ses-{session}][/{datatype}]/sub-{subject}'
-            '[_ses-{session}][_acq-{acquisition}][_run-{run}]_{suffix}.nii.gz", and the different fields will be '
-            "substituted based on the structure of bids_dir."
-        ),
-        type=str,
-        default=MASK_PATTERN,
-    )
-
-    parser.add_argument(
-        "--seed",
-        type=int,
-        default=42,
-        help="Seed to control the randomization (to be used with randomize=True).",
-    )
-
-    parser.add_argument(
-        "--classification",
-        help="Whether to perform classification.",
-        action="store_true",
-        default=True,
-    )
-    parser.add_argument(
-        "--regression",
-        help="Whether to perform regression.",
-        dest="classification",
-        action="store_false",
-    )
-    parser.add_argument(
-        "--custom_model",
-        help="Path to a custom model, trained using run_train_fetmrqc.py.",
-        default=None,
-        type=str,
-    )
-
-
 def check_fixed_args(fixed_args: dict) -> None:
     """Check that the fixed arguments equal to their defaults.
     This is because not all arguments can be easily changed from the command line without
     being mounted explicitly on the docker.
     """
-    defaults_dict = {"ckpt_path": None, "custom_model": None}
+    defaults_dict = {"ckpt_path": [None, BRAIN_CKPT], "custom_model": [None]}
     for k, v in fixed_args.items():
-        if v != defaults_dict[k]:
-            raise ValueError(
+        if v not in defaults_dict[k]:
+            raise Warning(
                 f"Argument {k} cannot be changed yet using the run_docker script. \n"
-                "It is fixed to {defaults_dict[k]}, but was passed as {v}."
+                f"It is fixed to {defaults_dict[k]}, but was passed as {v}. It is ignored."
             )
 
 
-def check_paths(out_dict: dict) -> None:
+def check_paths(out_dict: dict, out_dir) -> None:
     """Check that the paths that should be part of the output are given
     as relative paths. Otherwise, they cannot be mounted on the docker.
     """
@@ -212,7 +54,7 @@ def check_paths(out_dict: dict) -> None:
     for k, path in out_dict.items():
         if os.path.isabs(path):
             raise ValueError(
-                f"Output paths should be relative to the out_dir, but the argument {k} was set to {path}."
+                f"Output paths ({out_dict.keys()}) should be relative paths with respect to the out_dir ({out_dir}), but the argument {k} was set to {path}."
             )
 
 
@@ -221,15 +63,17 @@ def run_reports(args: argparse.Namespace) -> None:
     Run the reports pipeline using the given arguments.
     """
     # Create output directory
+    reports_dir = "reports" if args.reports_dir is None else args.reports_dir
     os.makedirs(args.out_dir, exist_ok=True)
-    os.makedirs(os.path.join(args.out_dir, args.reports_dir), exist_ok=True)
+    os.makedirs(os.path.join(args.out_dir, reports_dir), exist_ok=True)
     os.makedirs(
         os.path.dirname(os.path.join(args.out_dir, args.bids_csv)),
         exist_ok=True,
     )
     # Set paths
     out_docker = "/data/out"
-    reports_dir = os.path.join(out_docker, args.reports_dir)
+    reports_dir_docker = os.path.join(out_docker, reports_dir)
+
     bids_csv = os.path.join(out_docker, args.bids_csv)
     bids_dir = os.path.abspath(args.bids_dir)
     masks_dir = os.path.abspath(args.masks_dir)
@@ -239,7 +83,7 @@ def run_reports(args: argparse.Namespace) -> None:
 
     cmd = (
         f"docker run --rm -it "
-        "--gpus all --gpus all --ipc=host "  # Recommended by NVIDIA
+        "--gpus all --ipc=host "  # Recommended by NVIDIA
         "--ulimit memlock=-1 --ulimit stack=67108864 "  # Recommended by NVIDIA
         f"-v {bids_dir}:/data/data "
         f"-v {masks_dir}:/data/masks "
@@ -247,7 +91,7 @@ def run_reports(args: argparse.Namespace) -> None:
         f"{args.docker_path} qc_reports_pipeline "
         f"--bids_dir /data/data "
         f"--masks_dir /data/masks "
-        f"--reports_dir {reports_dir} "
+        f"--reports_dir {reports_dir_docker} "
         f"--mask_pattern {args.mask_pattern} "
         f"--bids_csv {bids_csv} "
         f"--seed {args.seed} "
@@ -311,6 +155,7 @@ def run_inference(args: argparse.Namespace) -> None:
         f"{fetmrqc20} {class_regr} "
         f"--mask_pattern {args.mask_pattern} "
         f"--seed {args.seed} "
+        f"--device {args.device} "
     )
     # Run command
     print(f"Running command: {cmd}")
@@ -369,6 +214,8 @@ def main() -> None:
             "reports_dir": args.reports_dir,
             "bids_csv": args.bids_csv,
         }
+        if args.reports_dir is None:
+            outputs_check.pop("reports_dir", None)
 
     elif args.command == "inference":
         outputs_check = {
@@ -383,9 +230,9 @@ def main() -> None:
         )
     else:
         raise ValueError(f"Unknown command {args.command}")
-
+    print(outputs_check)
     check_fixed_args(fixed_args)
-    check_paths(outputs_check)
+    check_paths(outputs_check, args.out_dir)
 
     if args.command == "reports":
         run_reports(args)
